UX polish: light theme, per-faction colours, wider sliders, factions dropdown, 200x200 default
- Default speed: max (0ms), default grid: 200x200
- Faction colours always shown per-faction regardless of unique-matrix toggle
- Colour palette reordered: red, blue, yellow, green, purple, orange, pink, teal, indigo, black
- Factions control changed from slider to dropdown (2–10 factions)
- Speed slider now fills available width
- Movement matrix section separated from colour cards
- Uint16Array for attackedBy bitmask to support up to 10 factions

+++ spiral/js/engine.js
@@ -31,7 +31,7 @@ export class SimEngine {
     // occupiedBy[i]: 0 = free, 1..6 = factionIndex+1
     this.occupiedBy = new Uint8Array(this.n);
     // attackedBy[i]: bitmask -- bit f means faction f attacks this cell
-    this.attackedBy = new Uint8Array(this.n);
+    this.attackedBy = new Uint16Array(this.n);
     // Per-faction scan pointers (only advance, never retreat -- O(n) total)
     this.scanPtrs = new Int32Array(this.factionCount);
     this.turn = 0;
